+++ mopidy/mpd/protocol/playback.py
@@ -42,7 +42,8 @@ def mixrampdb(context, decibels):
     doesn't fade the track volume, just overlaps. The songs need to have
     MixRamp tags added by an external tool. 0dB is the normalized maximum
     volume so use negative values, I prefer -17dB. In the absence of mixramp
-    tags crossfading will be used. See http://sourceforge.net/projects/mixramp
+    tags crossfading will be used. See
+    https://sourceforge.net/projects/mixramp/
     """
     raise exceptions.MpdNotImplemented  # TODO